+++ pyresttest/resttest.py
@@ -13,17 +13,21 @@ from email import message_from_string  # For headers handling
 import time
 
 try:
-    from cStringIO import StringIO
+    from cStringIO import StringIO as MyIO
 except:
     try:
-        from StringIO import StringIO
+        from StringIO import StringIO as MyIO
     except ImportError:
-        from io import StringIO
+        from io import BytesIO as MyIO
 
+ESCAPE_DECODING = 'string-escape'
 # Python 3 compatibility
-if sys.version_info[0] == 3:
+if sys.version_info[0] > 2:
     from past.builtins import basestring
     from builtins import range as xrange
+    ESCAPE_DECODING = 'unicode_escape'
+
+from six import text_type
 
 # Pyresttest internals
 from binding import Context
@@ -34,6 +38,7 @@ import validators
 from validators import Failure
 from tests import Test, DEFAULT_TIMEOUT
 from benchmarks import Benchmark, AGGREGATES, METRICS, parse_benchmark
+
 """
 Executable class, ties everything together into the framework.
 Module responsibilities:
@@ -43,7 +48,7 @@ Module responsibilities:
 - Collect and report on test/benchmark results
 - Perform analysis on benchmark results
 """
-
+HEADER_ENCODING ='ISO-8859-1' # Per RFC 2616
 LOGGING_LEVELS = {'debug': logging.DEBUG,
                   'info': logging.INFO,
                   'warning': logging.WARNING,
@@ -139,9 +144,6 @@ class TestResponse:
     def __str__(self):
         return json.dumps(self, default=safe_to_json)
 
-    def unicode_body(self):
-        return unicode(self.body.decode('UTF-8'))
-
 
 def read_test_file(path):
     """ Read test file at 'path' in YAML """
@@ -155,6 +157,8 @@ def parse_headers(header_string):
     """ Parse a header-string into individual headers
         Implementation based on: http://stackoverflow.com/a/5955949/95122
         Note that headers are a list of (key, value) since duplicate headers are allowed
+
+        NEW NOTE: keys & values are unicode strings, but can only contain ISO-8859-1 characters
     """
     # First line is request line, strip it out
     if not header_string:
@@ -162,6 +166,12 @@ def parse_headers(header_string):
     request, headers = header_string.split('\r\n', 1)
     if not headers:
         return list()
+
+    # Python 2.6 message header parsing fails for Unicode strings, 2.7 is fine. Go figure.
+    if sys.version_info < (2,7):
+        header_msg = message_from_string(headers.encode(HEADER_ENCODING))
+        return [(text_type(k.lower(), HEADER_ENCODING), text_type(v, HEADER_ENCODING))
+            for k, v in header_msg.items()]
     else:
         header_msg = message_from_string(headers)
         # Note: HTTP headers are *case-insensitive* per RFC 2616
@@ -211,7 +221,7 @@ def parse_testsets(base_url, test_structure, test_files=set(), working_directory
                 elif key == u'url':  # Simple test, just a GET to a URL
                     mytest = Test()
                     val = node[key]
-                    assert isinstance(val, str) or isinstance(val, unicode)
+                    assert isinstance(val, basestring)
                     mytest.url = base_url + val
                     tests_out.append(mytest)
                 elif key == u'test':  # Complex test with additional parameters
@@ -287,8 +297,8 @@ def run_test(mytest, test_config=TestConfig(), context=None):
     result.test = templated_test
 
     # reset the body, it holds values from previous runs otherwise
-    headers = StringIO()
-    body = StringIO()
+    headers = MyIO()
+    body = MyIO()
     curl.setopt(pycurl.WRITEFUNCTION, body.write)
     curl.setopt(pycurl.HEADERFUNCTION, headers.write)
     if test_config.verbose:
@@ -330,7 +340,7 @@ def run_test(mytest, test_config=TestConfig(), context=None):
     # Retrieve values
     result.body = body.getvalue()
     body.close()
-    result.response_headers = headers.getvalue()
+    result.response_headers = text_type(headers.getvalue(), HEADER_ENCODING)  # Per RFC 2616
     headers.close()
 
     response_code = curl.getinfo(pycurl.RESPONSE_CODE)
@@ -393,7 +403,7 @@ def run_test(mytest, test_config=TestConfig(), context=None):
     if test_config.print_bodies or not result.passed:
         if test_config.interactive:
             print("RESPONSE:")
-        print(result.body.decode("string-escape"))
+        print(result.body.decode(ESCAPE_DECODING))
 
     if test_config.print_headers or not result.passed:
         if test_config.interactive: