+++ pyresttest/test_parsing.py
@@ -1,9 +1,39 @@
+# -*- coding: utf-8 -*-
 import unittest
-from parsing import flatten_dictionaries, lowercase_keys, safe_to_bool, safe_to_json
+from parsing import *
 
+PYTHON_MAJOR_VERSION = sys.version_info[0]
 
 class TestParsing(unittest.TestCase):
     """ Tests for parsing utility functions """
+
+    def test_encode_unicode_bytes(self):
+        val = 8
+        unicoded = u'指事字'
+        byteform = b'\xe6\x8c\x87\xe4\xba\x8b\xe5\xad\x97'
+        num = 156
+
+        self.assertEqual(byteform, encode_unicode_bytes(unicoded))
+        self.assertEqual(byteform, encode_unicode_bytes(byteform))
+        self.assertEqual(b'156', encode_unicode_bytes(num))
+
+    def test_unicode_templating(self):
+        # Unicode template and unicode substitution
+        unicode_template_string = u'my name is 指 and my value is $var'
+        unicode_variables = {'var': u'漢'}
+        normal_variables = {'var': u'bob'}
+        substituted = safe_substitute_unicode_template(unicode_template_string, unicode_variables)
+        self.assertEqual(u'my name is 指 and my value is 漢', substituted)
+
+        # Normal template and unicode substitution
+        normal_template_string = 'my normal name is blah and my unicode name is $var'
+        substituted = safe_substitute_unicode_template(normal_template_string, unicode_variables)
+        self.assertEqual(u'my normal name is blah and my unicode name is 漢', substituted)
+
+        # Unicode template and normal substitution
+        substituted = safe_substitute_unicode_template(unicode_template_string, normal_variables)
+        self.assertEqual(u'my name is 指 and my value is bob', substituted)
+
 
     def test_flatten(self):
         """ Test flattening of lists of dictionaries to single dictionaries """
@@ -77,8 +107,12 @@ class TestParsing(unittest.TestCase):
             pass  # Good
 
     def test_safe_to_json(self):
-        self.assertEqual(u'adj12321nv', safe_to_json(
-            bytearray('adj12321nv', 'UTF-8')))
+        if PYTHON_MAJOR_VERSION == 2:
+            self.assertEqual(u'adj12321nv', safe_to_json(
+                bytearray(b'adj12321nv', 'UTF-8')))
+        else:
+            self.assertEqual(u'adj12321nv', safe_to_json(u'adj12321nv'))
+
         self.assertEqual(u'5.2', safe_to_json(5.2))
 
         class Special(object):